+++ src/app/table/data-table/cell/terra-data-table-header-cell.interface.ts
@@ -1,7 +1,7 @@
 /**
  * @author mkunze
  */
-import { TerraTextAlignEnum } from './terra-text-align.interface';
+import { TerraTextAlignEnum } from './terra-text-align.enum';
 
 export interface TerraDataTableHeaderCellInterface
 {
