+++ src/app/table/data-table/cell/terra-text-align.enum.ts
